2005_01_06 - hehe2 : (sequencer)
(*) _ Fix the selection algorithm to select patterns that have any part
in the selection zone (previously, they had to be entirely included in it).
(+) _ Added the ability to add patterns with a selection zone using the shift
key to the selected patterns list.
(+) _ Added the ability to drag multiple patterns previously selected together.


git-svn-id: https://svn.code.sf.net/p/wired/code/trunk/wired@95 6f5cd6f6-390f-0410-abd1-b47fe60d5253