2005_02_09 - hehe2 : (sequencer)
(*) _ Removed the inter tracks wxStaticLine objects and replaced them by a
drawing on the dc. It's faster and now there's not the 'limited width' bug.
(*) _ Fixed the splitting code, it's now fully working for audio and working
most of the time for midi patterns (they're still a bug to sort out on some
odd midifiles)
(*) _ Fixed the copy / paste on midi patterns (now it's working)
(*) _ Fixed some cosmetic bugs and cleaned some code (far from finished ...)


git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@196 6f5cd6f6-390f-0410-abd1-b47fe60d5253